+++ danling/models/transformer/attention/multihead_attention.py
@@ -2,9 +2,8 @@ import warnings
 from typing import Any, Dict, List, Optional, Tuple
 
 import torch
-import torch.nn as nn
 import torch.nn.functional as F
-from torch import Tensor
+from torch import Tensor, nn
 
 
 class MultiHeadAttention(nn.Module):
