+++ danling/models/transformer/encoder.py
@@ -2,9 +2,7 @@ from functools import partial
 from typing import Any, Dict, Optional
 
 import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch import Tensor
+from torch import Tensor, nn
 from torch.utils.checkpoint import checkpoint
 
 from .attention import MultiHeadAttention
