+++ danling/models/transformer/pos_embed/pos_embed.py
@@ -2,8 +2,7 @@ import math
 from typing import Optional
 
 import torch
-import torch.nn as nn
-from torch import Tensor
+from torch import Tensor, nn
 
 
 # this is from T5
