+++ danling/optim/lr_scheduler/lr_scheduler.py
@@ -1,4 +1,3 @@
-import math
 from typing import Optional
 
 import numpy as np
