+++ danling/utils/__init__.py
@@ -1,5 +1,4 @@
-from .decorator import flexible_decorator
-from .exception import catch
+from .decorator import catch, flexible_decorator
 from .io import is_json_serializable, load
 
 __all__ = ["catch", "flexible_decorator", "load", "is_json_serializable"]
